Pokemon Java Battle Simulation BETA version 1.3.5 release notes:
-Fixed damage calculation logical error
-Fixed typo for Sceptile
-Changed formatting for Pokemon selection and in-battle info
-Changed formatting for multi-type Pokemon
-Changed wording for two player menu
-Minor code cleanup

+++ Pokemon/src/battle/BattleEngine.java
@@ -514,9 +514,9 @@ public class BattleEngine {
 		}
 		else
 			STAB = move.getType() == pokemon[atk].getType() ? 1.5 : 1.0;
-		
+
 		// damage formula reference: https://bulbapedia.bulbagarden.net/wiki/Damage (GEN IV)
-		int damageDealt = (int)((Math.floor(((((Math.floor((2 * level) / 5)) + 2) * power * (A / D)) / 50)) + 2) * crit * STAB * type);
+		int damageDealt = (int)((Math.floor(((((Math.floor((2 * level) / 5)) + 2) * power * (int)(A / D)) / 50)) + 2) * crit * STAB * type);
 
 		// keep damage dealt less than or equal to remaining HP
 		if (damageDealt > pokemon[trg].getHP())
